elfnote: always use the "a" flag for .note.* sections
GNU as allows specifying different flags when constructing sections, for
example when the second occurrence has empty flags. But clang may not be so
permissive:
    $ cat test.s
    .pushsection .note.Linux, "a", @note ;
    .pushsection .note.Linux, "", @note ;
    $ as test.s
    $ clang-11 -cc1as -triple x86_64-pc-linux-gnu test.s
    test.s:2:1: error: changed section flags for .note.Linux

The kernel reaches this case when both ELFNOTE_START and ELFNOTE add data
to the same section, because the former is used exclusively with the "a"
flag and the latter always uses empty flags.

In vmlinux, the combined .notes section is marked as SHF_ALLOC, so use
the "a" flag for all these macros.

Link: ClangBuiltLinux#913
Signed-off-by: Ilie Halip <ilie.halip@gmail.com>